Add HeightMapProfiler.
This is a simple application for benchmarking 3D performance of Android handsets.  It generates a height map-based landscape and allows the user to tweak OpenGL ES state and scene complexity to see how various Android devices perform.  When the test is complete, profiling results (including average verts per frame and total frame time) are reported.

git-svn-id: https://apps-for-android.googlecode.com/svn/trunk@169 4c97771b-2a47-0410-a901-cdf36a16b7d0